% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCSIS.R
\name{Cor}
\alias{Cor}
\title{Parallel function
This is a parallel function about the projection correlation.}
\usage{
Cor(Xj, A_y, n)
}
\arguments{
\item{Xj}{Each column from design matrix of dimensions n * p}

\item{A_y}{The arccos value about Y}

\item{n}{The sample size}
}
\value{
the projection correlation between Xj and A_y
}
\description{
Parallel function
This is a parallel function about the projection correlation.
}
