% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GendataMRM.R
\name{GendataMRM}
\alias{GendataMRM}
\title{Generate simulation data (Multivariate response models)}
\usage{
GendataMRM(n, p, rho, type = c("a", "b"))
}
\arguments{
\item{n}{Number of subjects in the dataset to be simulated. It will also equal to the
number of rows in the dataset to be simulated, because it is assumed that each
row represents a different independent and identically distributed subject.}

\item{p}{Number of predictor variables (covariates) in the simulated dataset.
These covariates will be the features screened by model-free procedures.}

\item{rho}{The correlation between adjacent covariates in the simulated matrix X.
The within-subject covariance matrix of X is assumed to has the same form as an
AR(1) auto-regressive covariance matrix, although this is not meant to imply
that the X covariates for each subject are in fact a time series. Instead, it is just
used as an example of a parsimonious but nontrivial covariance structure. If
rho is left at the default of zero, the X covariates will be independent and the
simulation will run faster.}

\item{type}{The type of multivariate response models, which use different mean and covariance
structure to generate data. Specially, type="a" is following the Model 3.a and
type="b" is following the Model 3.b by Li et al.(2020).}
}
\value{
the list of your simulation data
}
\description{
This function helps you quickly generate simulation data based on transformation model.
You just need to input the sample and dimension of the data
you want to generate and the covariance parameter rho.
This simulated example comes from Example 3 introduced by Li et al.(2020)
}
\examples{
n <- 100
p <- 200
rho <- 0.5
data <- GendataMRM(n, p, rho, type = "a")

}
\references{
Liu, W., Y. Ke, J. Liu, and R. Li (2020). Model-free feature screening and FDR control with knockoff features. Journal of the American Statistical Association, 1–16.
}
\author{
Xuewei Cheng \email{xwcheng@hunnu.edu.cn}
}
