% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCSIS.R
\name{get_arccos}
\alias{get_arccos}
\title{Arccos function}
\usage{
get_arccos(X)
}
\arguments{
\item{X}{The design matrix of dimensions n * p. Each row is an observation vector.}
}
\value{
the arccos value
}
\description{
This is a function to get a arccos value based on projection correlation from the Python language.
}
