% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCSIS.R
\name{projection_corr}
\alias{projection_corr}
\title{Projection correlation function}
\usage{
projection_corr(A_x, A_y, n)
}
\arguments{
\item{A_x}{The arccos value about X}

\item{A_y}{The arccos value about Y}

\item{n}{The sample size}
}
\value{
the projection correlation
}
\description{
Projection correlation between X[,j] and Y from the Python language
}
